Update from Retail_Dev
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C33583

+++ Retail/Code/Retail/Retail/Tables/Account.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [Retail].[Account] (
-    [ID]          INT            IDENTITY (1, 1) NOT NULL,
+    [ID]          BIGINT         IDENTITY (1, 1) NOT NULL,
     [Name]        NVARCHAR (255) NOT NULL,
     [Type]        INT            NOT NULL,
     [Settings]    NVARCHAR (MAX) NULL,
@@ -8,4 +8,6 @@
     [timestamp]   ROWVERSION     NULL,
     CONSTRAINT [PK_Account] PRIMARY KEY CLUSTERED ([ID] ASC)
 );
+
+
